+++ packages/at_secondary_server/test/sync_verb_test.dart
@@ -15,11 +15,9 @@ import 'package:mocktail/mocktail.dart';
 
 class MockSecondaryKeyStore extends Mock implements SecondaryKeyStore {}
 
-class MockByteBuffer extends Mock implements ByteBuffer {}
-
 void main() {
   SecondaryKeyStore mockKeyStore = MockSecondaryKeyStore();
-  ByteBuffer mockByteBuffer = MockByteBuffer();
+
   group('A group of sync verb regex test', () {
     test('test sync correct syntax', () {
       var verb = Sync();
@@ -138,34 +136,32 @@ void main() {
         .thenAnswer((invocation) => Future(() => AtData()));
 
     test('test to ensure at least one entry is synced always', () async {
-      when(() => mockByteBuffer.isOverFlow(any())).thenReturn(true);
-
       verbHandler = SyncProgressiveVerbHandler(mockKeyStore);
-      var syncResponse = [];
-      var atCommitLog =
-          await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice'));
+      AtCommitLog atCommitLog = (await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice')))!;
 
       // Creating dummy commit entries
-      await atCommitLog?.commit('test_key_alpha', CommitOp.UPDATE_ALL);
-      await atCommitLog?.commit('test_key2_beta', CommitOp.UPDATE);
+      await atCommitLog.commit('test_key_alpha', CommitOp.UPDATE_ALL);
+      await atCommitLog.commit('test_key2_beta', CommitOp.UPDATE);
       // ensure commitLog is not empty
-      assert(atCommitLog!.entriesCount() > 0);
+      assert(atCommitLog.entriesCount() > 0);
 
-      await verbHandler.populateSyncBuffer(
-          mockByteBuffer, syncResponse, atCommitLog!.getEntries(0));
-
-      expect(syncResponse, isNotEmpty);
+      List<KeyStoreEntry> syncResponse = [];
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(0));
+      expect(syncResponse.length, 1);
+      expect(syncResponse[0].key, 'test_key_alpha');
     });
 
-    test(
-        'overflowing entry not added to syncResponse when syncResponse not empty',
-        () async {
-      when(() => mockByteBuffer.isOverFlow(any())).thenReturn(true);
-
+    test('overflowing entry not added to syncResponse when syncResponse not empty', () async {
       verbHandler = SyncProgressiveVerbHandler(mockKeyStore);
-      var syncResponse = [];
-      var atCommitLog =
-          await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice'));
+      AtCommitLog atCommitLog = (await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice')))!;
+
+      List<KeyStoreEntry> syncResponse = [];
+
+      // Creating dummy commit entries
+      await atCommitLog.commit('test_key_alpha', CommitOp.UPDATE_ALL);
+      await atCommitLog.commit('test_key2_beta', CommitOp.UPDATE);
+      // Ensure commitLog is not empty
+      expect(atCommitLog.entriesCount(), greaterThan(0));
 
       var entry = KeyStoreEntry()
         ..key = 'dummy'
@@ -175,70 +171,82 @@ void main() {
       // Inserting an element into syncResponse, so that now it isn't empty
       syncResponse.add(entry);
 
-      // Creating dummy commit entries
-      await atCommitLog?.commit('test_key_alpha', CommitOp.UPDATE_ALL);
-      await atCommitLog?.commit('test_key2_beta', CommitOp.UPDATE);
-      // Ensure commitLog is not empty
-      assert(atCommitLog!.entriesCount() > 0);
-
-      // Since syncResponse already has an entry, the next overflowing entry
+      // Since syncResponse already has an entry, and the 'capacity' is 0, then the next entry
       // should not be added to the syncResponse
-      await verbHandler.populateSyncBuffer(
-          mockByteBuffer, syncResponse, atCommitLog!.getEntries(1));
-
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(0));
       expect(syncResponse, [entry]);
+
+      syncResponse.clear();
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(0));
+      expect(syncResponse.length, 1);
+      expect(syncResponse[0].key, 'test_key_alpha');
+
+      syncResponse.clear();
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(1));
+      expect(syncResponse.length, 1);
+      expect(syncResponse[0].key, 'test_key2_beta');
     });
 
-    test('test to ensure all entries are synced if buffer does not overflow',
-        () async {
-      when(() => mockByteBuffer.isOverFlow(any())).thenReturn(false);
-
+    test('test to ensure all entries are synced if buffer does not overflow', () async {
       verbHandler = SyncProgressiveVerbHandler(mockKeyStore);
-      var syncResponse = [];
-      var atCommitLog =
-          await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice'));
-      mockByteBuffer.capacity = 1000000;
+      AtCommitLog atCommitLog = (await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice')))!;
 
       // Creating dummy commit entries
-      await atCommitLog?.commit('test_key_alpha', CommitOp.UPDATE_ALL);
-      await atCommitLog?.commit('test_key2_beta', CommitOp.UPDATE);
-      await atCommitLog?.commit('abcd', CommitOp.UPDATE_ALL);
-      await atCommitLog?.commit('another_random_key', CommitOp.UPDATE_META);
-      // ensure commitLog is not empty
-      assert(atCommitLog!.entriesCount() > 0);
+      await atCommitLog.commit('test_key_alpha', CommitOp.UPDATE_ALL);
+      await atCommitLog.commit('test_key2_beta', CommitOp.UPDATE);
+      await atCommitLog.commit('abcd', CommitOp.UPDATE_ALL);
+      await atCommitLog.commit('another_random_key', CommitOp.UPDATE_META);
 
-      await verbHandler.populateSyncBuffer(
-          mockByteBuffer, syncResponse, atCommitLog!.getEntries(0));
+      // ensure commitLog is not empty
+      var commitLogLength = atCommitLog.entriesCount();
+      expect(commitLogLength, 4);
+
+      List<KeyStoreEntry> syncResponse = [];
+      var entry = KeyStoreEntry()
+        ..key = 'dummy'
+        ..commitId = 11
+        ..operation = CommitOp.UPDATE_ALL
+        ..value = 'whatever';
+      // Inserting an element into syncResponse, so that now it isn't empty
+      syncResponse.add(entry);
+
+      await verbHandler.prepareResponse(10 * 1024 * 1024, syncResponse, atCommitLog.getEntries(0));
 
       // Expecting that all the entries in the commitLog have been
       // added to syncResponse
-      expect(syncResponse.length, atCommitLog.entriesCount());
+      expect(syncResponse.length, commitLogLength + 1);
+      expect(syncResponse[0], entry);
+      expect(syncResponse[1].key, 'test_key_alpha');
+      expect(syncResponse[2].key, 'test_key2_beta');
+      expect(syncResponse[3].key, 'abcd');
+      expect(syncResponse[4].key, 'another_random_key');
     });
 
-    test('ensure only one overflowing entry is added to syncResponse'
-        ' when commitLog has two large entries',
-            () async {
-          when(() => mockByteBuffer.isOverFlow(any())).thenReturn(true);
+    test('ensure only one overflowing entry is added to syncResponse when commitLog has two large entries', () async {
+      verbHandler = SyncProgressiveVerbHandler(mockKeyStore);
+      AtCommitLog atCommitLog = (await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice')))!;
 
-          verbHandler = SyncProgressiveVerbHandler(mockKeyStore);
-          var syncResponse = [];
-          var atCommitLog =
-          await (AtCommitLogManagerImpl.getInstance().getCommitLog('@alice'));
-          mockByteBuffer.capacity = 1000000;
+      // Creating dummy commit entries
+      await atCommitLog.commit('test_key1', CommitOp.UPDATE_ALL);
+      await atCommitLog.commit('test_key2', CommitOp.UPDATE);
+      // ensure commitLog is not empty
+      assert(atCommitLog.entriesCount() == 2);
 
-          // Creating dummy commit entries
-          await atCommitLog?.commit('test_key1', CommitOp.UPDATE_ALL);
-          await atCommitLog?.commit('test_key2', CommitOp.UPDATE);
-          // ensure commitLog is not empty
-          assert(atCommitLog!.entriesCount() > 0);
+      List<KeyStoreEntry> syncResponse = [];
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(0));
+      expect(syncResponse.length, 1);
+      expect(syncResponse[0].key, 'test_key1');
 
-          await verbHandler.populateSyncBuffer(
-              mockByteBuffer, syncResponse, atCommitLog!.getEntries(0));
+      syncResponse.clear();
+      await verbHandler.prepareResponse(0, syncResponse, atCommitLog.getEntries(1));
+      expect(syncResponse.length, 1);
+      expect(syncResponse[0].key, 'test_key2');
 
-          // Expecting that all the entries in the commitLog have been
-          // added to syncResponse
-          expect(syncResponse.length, 1);
-        });
+      // test with empty iterator
+      syncResponse.clear();
+      await verbHandler.prepareResponse(10 * 1024 * 1024, syncResponse, atCommitLog.getEntries(2));
+      expect(syncResponse.length, 0);
+    });
 
     tearDown(() async => await tearDownFunc());
   });